% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_div.R
\name{beta_boxplot}
\alias{beta_boxplot}
\title{Within beta-diversity boxplot}
\usage{
beta_boxplot(physeq, method = "bray", group)
}
\arguments{
\item{physeq}{phyloseq-class object.}

\item{method}{beta-diversity metric. Default "bray", i.e., Bray-Curtis dissimilarity.}

\item{group}{factorial variable to group.}
}
\value{
A list with the data frame and with the box plot of within group beta-diversity.
}
\description{
Given a phyloseq object, a beta-diversity metric and a group/factorial variable
to group the samples, it calculates the within group beta-diversity.
}
