% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax.R
\name{filter_feature_table}
\alias{filter_feature_table}
\title{Filter phyloseq object based on samples and taxa.}
\usage{
filter_feature_table(
  physeq,
  taxa2filter = NULL,
  samples2filter = NULL,
  min_sample_depth = 0,
  min_taxa_counts = 0
)
}
\arguments{
\item{physeq}{phyloseq-class object.}

\item{taxa2filter}{taxonomic names to exclude (character). It can include the
taxa names, i.e., OTU/ASV ids - 'taxa_names(physeq)'. Also it can be instead a logical
of the same length as OTU/ASV ids (i.e., 'length(taxa_names(physeq))'), where 'TRUE'
represents the taxa to keep and 'FALSE' to remove. Default 'NULL'.}

\item{samples2filter}{name of the samples to discard (character). Default 'NULL'.}

\item{min_sample_depth}{minimum number of reads per sample to keep a sample (numeric). Default '0'.}

\item{min_taxa_counts}{minimum number of reads per feature i.e., OTU/ASV, to keep
a feature (features sum across samples) (numeric). Default '0'.}
}
\value{
A phyloseq object filtered.
}
\description{
Filter phyloseq object based on samples and taxa names or minimum number of reads.
}
