% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{get_physeq_tbls}
\alias{get_physeq_tbls}
\title{Get phyloseq tables}
\usage{
get_physeq_tbls(physeq, tax_rank = NULL, rm_na = FALSE, feature_w_tax = FALSE)
}
\arguments{
\item{physeq}{phyloseq-class object with 'otu_table', 'tax_table', 'sam_data' slots (required).}

\item{tax_rank}{taxonomic rank (character). One of the taxonomic ranks among
the column names of 'tax_table()' of the 'physeq' object given.}

\item{rm_na}{include (TRUE) or not (FALSE) NAs (logical), i.e., taxa without classification at
the taxonomic level specified at 'tax_rank'. It only works if 'tax_rank' was provided.
Otherwise it is ignored.}

\item{feature_w_tax}{substitute the feature labels (in columns) by the taxa names specified at
'tax_rank' level (logical). Default is 'FALSE'. It only works if 'tax_rank' was provided.
Otherwise it is ignored.}
}
\value{
A list with the 'feature' ('otu_table'), 'taxonomy' ('tax_table'), and 'metadata'
('sample_data') tables. If the 'tax_rank' is specified, it will give 'feature' and 'taxonomy'
summarized at that 'tax_rank' level. If 'feature_w_tax' is equal 'TRUE' (only works together
with 'tax_rank'), the column names in the 'feature' table will be the taxa names at 'tax_rank'
level specied.
}
\description{
Given a phyloseq object with 'otu_table', 'tax_table', 'sam_data' (required)
it gives a list with those tables.
}
