% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_div.R
\name{good_coverage}
\alias{good_coverage}
\title{Good's coverage index}
\usage{
good_coverage(x)
}
\arguments{
\item{x}{numeric (of integers) matrix where rows are samples/observations and
columns are OTUs/ASVs/Species. If given a data frame or phyloseq-class object,
it will be coerced to a matrix.}
}
\value{
A named (with samples) vector of Good's coverage estimates per sample.
}
\description{
The 'good_coverage' function estimates the Good's coverage index
based on a matrix of OTUs/ASVs/Species counts.
}
