% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{melt_physeq}
\alias{melt_physeq}
\title{Melt phyloseq object}
\usage{
melt_physeq(physeq)
}
\arguments{
\item{physeq}{phyloseq-class object with 'otu_table', 'tax_table', 'sam_data' slots (required).}
}
\value{
A data frame with with all the OTU/ASV (tidy), taxonomic and sample data. The same
output to as 'psmelt()' function (from 'phyloseq').
}
\description{
Given a phyloseq object with 'otu_table', 'tax_table', 'sam_data' (required)
it gives the data.frame melted with all the OTU/ASV (tidy), taxonomic and sample data.
Same functionality of the function 'psmelt()' from the 'phyloseq' R package, but
fastest as it uses 'data.table' R package for the task.
}
