% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax.R
\name{plot_taxa_heatmap}
\alias{plot_taxa_heatmap}
\title{Plot heatmap}
\usage{
plot_taxa_heatmap(
  physeq,
  tax_rank = NULL,
  rm_na = FALSE,
  annot_samples = NULL,
  norm_mth = NULL,
  scale_by = NULL,
  tr_heat = FALSE,
  set_seed = 1024,
  ...
)
}
\arguments{
\item{physeq}{phyloseq-class object with 'otu_table', 'tax_table', 'sam_data' slots (required).}

\item{tax_rank}{taxonomic rank (character). One of the taxonomic ranks among
the column names of 'tax_table()' of the 'physeq' object given.}

\item{rm_na}{include ('TRUE') or not ('FALSE') NAs (logical), i.e., taxa without classification at
the taxonomic level specified at 'tax_rank'. It only works if 'tax_rank' was provided.
Otherwise it is ignored.}

\item{annot_samples}{one or more categorical/factor variables (character) among sample data
(in 'sample_variables(physeq)'). Default is 'NULL'.}

\item{norm_mth}{normalize/transform the feature table. Default is 'NULL'. One character to pass to
'vegan::decostand()' - one of c("total", "max", "frequency", "normalize", "range", "rank",
"standardize", "pa", "chi", "hellinger", "log").}

\item{scale_by}{scale (Z-score) data by "samples" or "taxa" (character). Default is 'NULL'.}

\item{tr_heat}{transpose heatmap. Default is 'FALSE' (logical).}

\item{set_seed}{set seed to allow reproducibility. Default is '1024' (numeric). Set to 'NULL' to
turn it off.}

\item{...}{parameters to pass to the function 'ComplexHeatmap::Heatmap()' with the exception of
'matrix'. Also 'top_annotation' (samples in columns) or 'left_annotation' (samples in rows)
cannot be specified if 'annot_samples' was specified.}
}
\value{
Heatmap (from 'ComplexHeatmap').
}
\description{
Plots an hierarchial heatmap with the package 'ComplexHeatmap' for a given a phyloseq object
with 'otu_table', 'tax_table', 'sam_data' (required). If provided a 'tax_rank' the heatmap
will be subsetted to the taxonomic rank specified. Samples can be annotated by specifying one
or more categorical/factor variables (character) among sample data (in 'sample_variables(physeq)').
}
