% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_div.R
\name{prevalence}
\alias{prevalence}
\title{Prevalence of features (OTUs/ASVs)}
\usage{
prevalence(physeq, tax_rank)
}
\arguments{
\item{physeq}{phyloseq-class object.}

\item{tax_rank}{taxonomic rank (character). One of the taxonomic ranks among
the column names of 'tax_table()' of the 'physeq' object given.}
}
\value{
A data frame with Feature_ID (lists the unique features (OTUs/ASVs)),
"Feature_[tax_rank]" (lists the taxonomic affiliation at the taxonomic rank defined
by 'tax_rank', Feature_Abundance (absolute abundance of each feature),
Feature_Relative (relative abundance of each feature), Prevalence (prevalence of
each feature - the no. of samples where each feature is observed),
Prevalence_Relative (the same as Prevalence, but in relative abundance).
}
\description{
Given a phyloseq object and a taxonomic rank (character) it gives the
prevalence of features (OTUs/ASVs), i.e., the number of samples where
each feature (OTUs/ASVs) appear (observed) grouped by taxonomic rank.
}
