% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax.R
\name{rank_taxa}
\alias{rank_taxa}
\title{Rank taxa}
\usage{
rank_taxa(physeq, tax_rank, top_taxa = NULL, type_count = "abs", ...)
}
\arguments{
\item{physeq}{phyloseq-class object.}

\item{tax_rank}{taxonomic rank (character). One of the taxonomic ranks among
the column names of 'tax_table()' of the 'physeq' object given.}

\item{top_taxa}{top most abundant taxa to select by 'tax_rank' (numeric - coerced
to integer). Default 'NULL'.}

\item{type_count}{type of counts to perform 'rank_taxa()'. It can be absolute, i.e.,
'abs', or percentage, i.e., 'perc' (character). Default is 'abs' - absolute counts.}
}
\value{
It returns a list with a tibble data frame ('$data') and ggplot ('$plot')
ranking the taxa picked at the taxonomic rank select at 'tax_rank'.
}
\description{
Given a phyloseq object (with absolute abundance counts) and a taxonomic
rank (character - one among 'rank_names(physeq)') it gives the taxa (by 'tax_rank')
summarized for the whole data set ranked by abundance.
}
