% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_div.R
\name{uni_taxa_per_sample}
\alias{uni_taxa_per_sample}
\title{Uquiness of features (OTUs/ASVs) per sample}
\usage{
uni_taxa_per_sample(physeq, tax_rank)
}
\arguments{
\item{physeq}{phyloseq-class object.}

\item{tax_rank}{taxonomic rank (character). One of the taxonomic ranks among
the column names of 'tax_table()' of the 'physeq' object given.}
}
\value{
A data frame with Taxa (lists the taxa that belongs to the
'tax_rank' provided), Sample (lists the samples), Count (no. of unique
features belonging to that taxa - 'Taxa' - and found in only that sample - 'Sample').
}
\description{
Given a phyloseq object and a taxonomic rank (character), it gives the
uniquess of features (OTUs/ASVs), i.e., the number of features
(OTUs/ASVs) observed in one sample only, per sample per taxa.
}
