% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_div.R
\name{uni_taxa_rank}
\alias{uni_taxa_rank}
\title{Uquiness of features (OTUs/ASVs)}
\usage{
uni_taxa_rank(physeq, tax_rank)
}
\arguments{
\item{physeq}{phyloseq-class object.}

\item{tax_rank}{taxonomic rank (character). One of the taxonomic ranks among
the column names of 'tax_table()' of the 'physeq' object given.}
}
\value{
A data frame with Taxa (lists the unique taxa that belongs to the
'tax_rank' provided), Unique_Taxa (no. of features - OTUs/ASVs - that appear
only once, in one sample), Total_Taxa (no. of distinct features),
Percentage_unique _Taxa (percentage of Unique_Taxa across the Total_Taxa).
}
\description{
Given a phyloseq object and a taxonomic rank (character) it gives the
uniquess of features (OTUs/ASVs), i.e., the number
of features (OTUs/ASVs) observed in one sample only.
}
